Create context api for cap expenditure
