+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "abiza-dashboard",
-  "version": "1.1.7",
+  "version": "1.1.8",
   "description": "Abiza Dashboard",
   "main": "index.js",
   "scripts": {
