+++ src/context/EarningsSectionContext.js
@@ -1,0 +1,3 @@
+import React, { createContext } from "react"
+
+export const EarningsSectionContext = createContext(null);
