+++ src/pages/Dashboard.scss
@@ -32,6 +32,7 @@
 
 .dashboardrowcolumn2 {
     grid-column: 2;
+    width: 100%
 }
 
 .dashboardbox {
